// src/app/shared/models/beer.ts
export interface Beer {
    product_id: number;
    name: string;
    size: string;
    price: number;
    beer_id: number;
    image_url: string;
    category: string;
    abv: number;
    style: string;
    attributes: string;
    type: string;
    brewer: string;
    country: string;
    on_sale: boolean;
}